Implemented proper use of UserDto
