+++ src/main/java/com/pnu/cs/timeout/dto/UserTransformer.java
@@ -1,0 +1,22 @@
+package com.pnu.cs.timeout.dto;
+
+import com.pnu.cs.timeout.model.User;
+
+public class UserTransformer {
+    public static User toEntity(UserDto dto) {
+        return User.builder()
+                .email(dto.getEmail())
+                .firstName(dto.getFirstName())
+                .lastName(dto.getLastName())
+                .password(dto.getPassword())
+                .build();
+    }
+
+    public static UserDto toDto(User user) {
+        return UserDto.builder()
+                .email(user.getEmail())
+                .firstName(user.getFirstName())
+                .lastName(user.getLastName())
+                .build();
+    }
+}
